FCN, README, sample output
