+++ cnn/samples/ang20200924t211102_ch4mf_v2y1_img_CalCh4_v8+COVID_QC+Permian_QC_AA_saliency_pad256.hdr
@@ -1,0 +1,16 @@
+ENVI
+description = {
+./ang20200924t211102_ch4mf_v2y1_img_CalCh4_v8+COVID_QC+Permian_QC_AA_saliency_pad256.img}
+samples = 669
+lines   = 2801
+bands   = 1
+header offset = 0
+file type = ENVI Standard
+data type = 4
+interleave = bsq
+byte order = 0
+map info = {UTM, 1, 1, 272247.152557, 3992010.65018, 3.1, 3.1, 11, North,WGS-84, rotation=17}
+coordinate system string = {PROJCS["unnamed",GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]],PROJECTION["Transverse_Mercator"],PARAMETER["False_Easting",500000.0],PARAMETER["False_Northing",0.0],PARAMETER["Central_Meridian",-117.0],PARAMETER["Scale_Factor",0.9996],PARAMETER["Latitude_Of_Origin",0.0],UNIT["Meter",1.0]]}
+band names = {
+Band 1}
+data ignore value = -9999
